restapi: Build correctly instance type NIC links
The collection of NICs of instance types has been implemented reusing
the class that implements the collection of NICs of templates. As a side
effect the links to these NICs point to the templates collection instead
of to the instance types collection. This patch introduces a new class
to implement the collection of NICs of instance types, so that the
"setParents" method that is used as part of the calculation of the link
can be overriden to set a instance type as parent instead of a template.

Change-Id: Ib3ab0857517a1fefc88b835457aa4b4241a3bb5d
Bug-Url: https://bugzilla.redhat.com/1241574
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>