builtin: internal: encode password as PBE and add defaults
this will enable to use the builtin internal extension as standalone
extension in more secure method.

will be useful for devenv and other special installation.

in future the EngineExtensionsManager will no longer load the internal
extension, setup will create extension configuration to load it, the
password can be encoded using the ovirt-engine-crypto-tool.

the engine-config AdminUser, AdminPassword setting will be depreciated once
above method will be applied.

for now, this change is transparent.

Change-Id: Icc206c58c867b82400fff54d2b46df5cc203111a
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>