webadmin: Enable force start gluster volume
This patch enables user to force start a degraded gluster
volume. The patch handles the following cases :
Case-1 : Start volumes selected from Volumes main tab
case-2 : Click on unknown status icon against each volume
It takes the following appraoach:
1. Enable user to start volume when some/all bricks of the
   volume(s) is down.
2. Separate out the list of volumes as
   2.1 Volumes that are down and don't need force start
   2.2 Volumes with status up but some/all bricks down.
       These volumes require force start as they are in
       degraded state.
3. Simply attempt start on volumes in list obtained from 2.1
4. Show force warning for volumes staisfying 2.2 (if any).
   On click of "OK" :
   4.1 If user selects force check box force start the volumes in
       the list.
   4.2 If force check box is not selected, attempt normal start
       with force = false.
Steps 3 and 4 are attempted with UI error-handling set to true.
This patch also relaxws a canDoAction to suit the requirement that
if the volume is "UP" but not all the bricks participating in the
volume are "UP", then force start needs to be permitted.

Change-Id: Ib421030714f190feea1737d8c97f2e6308e72199
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1182373
Signed-off-by: Anmol Babu <anbabu@redhat.com>