host-deploy: Support deploying Hosted Engine components
Till today a hosts with hosted engine components were added manually by
script on the hosts themselves. This patch will enable to install and
configure the relevant components by passing all the relevant
information from the engine.

The prerequisite for that is to have the hosted engine VM imported into
the engine. Otherwise we can't gather all the info for the installation.

So now a user can select to Deploy/Remove hosted engine agent or leave with no change(see Action)

In order to deploy an additional Hosted Engine host we need to provide
the configuration that was used by the 1st host and just assign it with
a differnt unique host_id. For that we download the configuration from a
special disk on the special domain and extract the configuration from
it. This behaviour is composed under the HostedEngineConfigFetcher.java
utility.

Wiki:
http://www.ovirt.org/develop/release-management/features/hosted-engine-add-hosts-with-web-ui/

Change-Id: I535bdfeba41b03c86520408e72abfa914fe3e1be
Bug-Url: https://bugzilla.redhat.com/1167262
Signed-off-by: Roy Golan <rgolan@redhat.com>
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>