core: Removed useless data memebers
The attachedDomains and isoMap members are constant empty maps. Since
they are never changed, there's no point in holding them as data
members. Instead, where they are used as mock return values, and inline
call to Collections.emptyList() is clearer, and leaves less room for
potential mistakes in the future.

Change-Id: I03d3b2109ef7939cf7ad7c77f6a7b0b214b3bf67
Signed-off-by: Allon Mureinik <amureini@redhat.com>