core: Add tests for isStorageDomainNotInAnyPool()
StorageDomainToPoolRelationValidartor#isStorageDomainNotInAnyPool()
used to be tested indirectly (read: wrongly) in
RemoveStorageDomainCommandTest.

This wrong behavior was removed in
a87cfdf6cedeaadb8fd3b69dd7567f581348414f, which, in spite of being
correct in its own right, had the unfortunate side effect of losing some
test coverage.
This patch introduces new tests for the methods we lost coverage on, in
the right place this time.

Change-Id: I7ed344b8e6e8e1a29f0417a5a8f5a7c82ff06086
Signed-off-by: Allon Mureinik <amureini@redhat.com>