core: fix options format message
Since fence-agents package removed support of sending boolean parameters
by their own (i.e. lanplus, ssl etc.) and now we are forced to send the
boolean value explicitly (i.e lanplus=1 ssl=1 etc.), we have to change
the help message in the UI for the options field to reflect that.

Change-Id: Iad4339a68608f29d279dce74fbd056c5dd80f2ad
Signed-off-by: emesika <emesika@redhat.com>