core: Resurrect VdsBrokerObjectBuilderTest
VdsBrokerObjectBuilderTest was commented out in patch
7469b8d5392840a30e1cb8e6a05eafce83114d6c, since it used PowerMock which
slowed down the build considerably.

This patch uncomments the class and properly removes PowerMock usages so
this test can be used again.
With regard to performance, this test runs in ~800 milliseconds on my
laptop, which is completely reasonable, although exact times may change
depending on the machine, the load it's experiencing, etc.

Note that this patch is pretty raw, and just introduces the minimal work
required to re-enable this test. Following patches in the series will
continue to improve this class as per the project's standards.

Changes performed:
* Uncommented all the class.
* Removed the top "TODO" comment, as this patch does exactly that.
* Removed PowerMockito calls and annotations.
* Replaced the test runner with MockitoJUnitRunner.
* Re-introduced JsonObjectDeserializer#deserializeUnformattedJson
  removed in b2983fad2002ac47cfc98a7d84401daa85874c5e that's needed for
  asserting the disk statistics results.
* Removed usages of XmlRpcStruct and replaced them with
  Map<String, Object> as done in VdsBrokerObjectsBuilder in commit
  fa3c96ad55ed3aac5b0a923c0a30363c57d5c293.
* Reordered the imports to comply with the checkstyle verification
  added in commit a4f50e3be1717fd9cc60314f87b3d56bae6ba048.

Change-Id: I7b05b287f7938fa26197dc4dfd3f5681ea7a86a2
Signed-off-by: Allon Mureinik <amureini@redhat.com>