core: Set VDS status to down after ssh shutdown
Fix that the state was only set to down for the VDS instance which was
passed into the command. When querying for the VDS in the database it
still showed that the VDS was up.

Further the audit log did not show that the ssh shutdown was executed.
Now it logs USER_VDS_STOP on success and USER_FAILED_VDS_STOP otherwise.

Change-Id: I7f837e5d179a3ed0090fa76eef965a7df673940d
Bug-Url: https://bugzilla.redhat.com/1062228
Signed-off-by: Roman Mohr <rmohr@redhat.com>