core: use db status when processing down vms
when analyzing vm in down status, if the vm is not running on that host,
the status from db is not taken in consideration.

in some cases, like during migration, the analyzer can use the status
from db, even if the vm is not running this host, in order to understand
the current state of the vm.

scenrio for example: vm moves to down on destination, since the analyzer
dont use the status from the db, it doesnt know its migration process
and clears the migrate command from the cache, breaking rerun/further
treatment of the command.

Change-Id: I1a893f7e06e417ca1a516cb04feb3a5defdb9c23
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>