engine : Fix brick state change messages on start volume
An attempt to sync the volume capacity details as part of
start volume was made. As part of this a new issue of
appearance of unwanted brick state change messages was observed.
The cause for this is attempting to sync volume advanced details
for the volume with status not updated. Hence, this patch fetches
the volume newly from db and then uses it to get advanced details.

Change-Id: I0c259601de0e8edfdb0ced91e582db3f31731f3e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1235560
Signed-off-by: Anmol Babu <anbabu@redhat.com>