core: get down vm from monitoring also on power off
currently power off flow assume vm will not be reported by vdsm anymore,
and set vm to down and call processDownVm directly.
this is not always true:
* if it takes some time to vdsm to destroy the vm
 the vm might be reported as down on monitoring cycle.
* when using events - event with status down in received.

so changed DestroyVmVDSCommand to set the vm to powering down, and down
will be received from the monitoring.
also changed the monitoring to allow vm in Powering Down status not to
be reported anymore, and treat as normal down (and not error).

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1242500
Change-Id: Ic60cceac6ad71adcb4d96a8938f98b5ea08f2043
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>