core: graphicsProtocol in changed fields list
After editing parameters of a running VM, list of changed fields that
will be updated on the next run is displayed to user. Graphics devices
are presented in this list as separate devices (SPICE and VNC) while
corresponding to the single field in the UI ("Graphics protocol").

To correct this, 'name' parameter was introduced into
@EditableDeviceOnVmStatusField annotation. If it is present, its value
will be displayed to the user in the changed fields list, instead of
device type name. If two devices have the same name, it will be included
into the list only once. This allows, in our case, to present SPICE and
VNC devices as single 'graphicsProtocol' field.

Change-Id: I95facdae456ca904f1e0d1ece32a96658992b2eb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1238742
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>