engine : Fix scenario where CoCo child executes Async Tasks
When CoCo command with callback executes a command that
starts an async task the endSuccessfully of the CoCo command
is being invoked as it is the parent of command that executes
the async task.

The CoCo command with callback should not be considered parent of
the command executing async task.

Change-Id: Ib3261800f383a6e3973acab4bd16079594fa4149
Bug-Url: https://bugzilla.redhat.com/1293644
Signed-off-by: Ravi Nori <rnori@redhat.com>