engine: Completed task still shown as 'running' in UI
Currently when SPM host goes down due to connectivity issue,
VdsNotRespondingTreatment command is left forever 'running'
(hourglass icon) in the WebAdmin tasks pane.

This happens becuase when SPM host goes down the datacenter
should go down, and thus an internal command:
SetStoragePoolStatus is invoked, and there is a mixup in
the Execution-Context.

This patch solves this bug locally, although a similar
issue may exist in other places and handling of
CommandContext and ExecutionContext should be revisited.

Bug-Url: https://bugzilla.redhat.com/1145099
Change-Id: I4f5fbaef3e8c57fbad7997cb67282a6c163b6b03
Signed-off-by: Ori Liel <oliel@redhat.com>