scheduling: Fix ConcurrentModificationException in runFilters()
Since the MultipleActionRunner invokes canDoActions in parallel,
in case of multiple RunVmCommands it caused the SchedulingManager#canSchedule()
method to call runFilters() concurrently which resulted in duplicate sort
of the same ArrayList in sortFilters() method.

Now RunFilters creates a local copy of the filters list and run sort on this
private copy.

Change-Id: I1725940c11c2cc07c20255c4474715020e36b56e
Signed-off-by: Martin Betak <mbetak@redhat.com>