backend: Fix Deadlock in multiple hostdev RunVmCommands
The eager synchronous instantiation of Commands in MultipleActionsRunner
combined with host device lock being obtained in @PostConstruct of RunVmCommand
resulted in deadlock in initCommandsAndReturnValues() method.

The deadlock was resolved by moving the acquireHostDeviceLock() to
RunVmCommand#canDoAction() and releasing it immediately after check
with cooperation of the runVm() method that obtains it again and holds
it during the actual host device operations.

Change-Id: I64135e46be0dfb6dd71dde862d44f5afa1f1a7b9
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1248104
Signed-off-by: Martin Betak <mbetak@redhat.com>