core: translate vm graphic settings in specParams
When Engine is sending the new-style graphic devices,
VDSM expects the parameters
'fileTransferEnable' and 'copyPasteEnable'
to be sent into the device specParams.

The backward compatibility on VDSM side works when
graphics devices aren't specified, and in this case
and in this case only the vm-level parameters are considered.

Until the UI settings will change directly the per-device
specParams, the best way forward is to translate the per-vm
settings into per-device specParams (format and naming is
the same).

Bug-Url: https://bugzilla.redhat.com/983019
Change-Id: I960bdee2f0828bd3e34f2e2ba8c169b8b83c8d9a
Signed-off-by: Francesco Romani <fromani@redhat.com>