webadmin: fix inaccurate error message
A storage domain that contains a disk of a vm which is in a state other
than "Down", cannot be taken down to maintenance.
If that's the situation, and one tries to take it down to maintenance,
he gets an error message telling him that "Active VMs were detected".
That is inaccurate, since this error message will arise in any vm state
other than "Down".
This patch fixes the error message.

Change-Id: I87e061d716f045fb1549ff8713eaa16d990c0928
Bug-Url: https://bugzilla.redhat.com/1061194
Signed-off-by: Idan Shaby <ishaby@redhat.com>