webadmin: fix network interfaces of imported vm using v2v
The network interfaces for a VM that is imported using virt-v2v were
not defined correctly. This patch fixes it by setting the network
name and profile properly.

In addition, a new column is added the the network interfaces sub-tab
of the import VM dialog in order to show the name of the network that
the interface is declared on in the remote environment.

Change-Id: I804e988c97cf72924c82d8bb1ca4c6cab4caeebb
Bug-Url: https://bugzilla.redhat.com/1236075
Signed-off-by: Arik Hadas <ahadas@redhat.com>