restapi: Remove unneeded calls in SD extend
Removing unneeded code when extending a storage domain.
The same logic already exists in ExtendSANStorageDomainCommand.

Now, the 'host' parameter is not needed and can be removed from the update
storage domain definition in the rsdl.

Backward compatibility is preserved as the 'host' parameter is not
removed from XML schema.
If the parameter is present, it will be ignored.

Change-Id: I45128d928421e27f9eac2b73ac6fc03f1682f545
Bug-Url: https://bugzilla.redhat.com/1213961
Signed-off-by: Fred Rolland <frolland@redhat.com>