core: fix network interfaces for imported vm using v2v
Add an optional field for network interface of VM that contains
the name of the network in the remote environment on which the
interface is declared. It is relevant for VMs that exist in
some external environment of course.

In addition, setting the name of network interface according to
the standard in oVirt, i.e nic1, nic2 and so on.

Change-Id: Icba912e15afd2f90799fac0c92b331740a05eb73
Bug-Url: https://bugzilla.redhat.com/1236075
Signed-off-by: Arik Hadas <ahadas@redhat.com>