HA reservation needs to take the full VM's memory into account
The current HA reservation code only counted the used memory of
HA VMs. But the engine will refuse to migrate the VM when there
is not enough memory when compared to the minimum guaranteed
settings.

This patch updates the computation and uses the minimum guaranteed
memory size or the currently used amount of memory whichever is
higher.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1220134
Change-Id: Ib3af7e370f004f7c22335c507e4e191b90166f6d
Signed-off-by: Martin Sivak <msivak@redhat.com>
(cherry picked from commit 4711d1fdf22265a6c970dc4b2fa3b2eb84646342)