core: Adding StorageJobCallback/StorageJobCommand
This patch adds StorageJobCommand which is supposed to be a base class
for the storage job initiating commands.

Additionally, this patch adds the StorageJobCallback which is supposed to be used by
BLL commands that creates a storage job on the vdsm side. The callback
tracks the created job and calls the needed methods in the command
according to the job status.

Change-Id: Ic8ceb9ec0ee9f74dcd33021ccf97d897bf2e8557
Signed-off-by: Liron Aravot <laravot@redhat.com>