core: CDA msg when removing a template's disk without specifing domain.
Deleting a template's disk must be performed with the disk's storage
domain, as deleting a template's disk is enabled only if this disk has
copies on other domains.

Hence, a storage domain must be provided when calling RemoveDiskCommand
with a template's disk.

This patch adds that condition into RemoveDiskCommand's CDA with a
proper message.

Bug-Url: https://bugzilla.redhat.com/1229212
Change-Id: Id5f0c1bda25b5fbadbf153bfcb84c78950216e66
Signed-off-by: Amit Aviram <aaviram@redhat.com>