Added dummy check-patch
Because not always will a job run for ovirt-engine, this ensures that
at least one job will run and the CI flag will be set.

The flow where this scripts comes into play, is when you submit a
patch, and before jenkins had time to give a review, you submit a new
patch that does no code changes (like just changing the commit
message). At that point, gerrit has no previous review to carry over
to the new patch and jenkins will not get triggered as it does not
have any code changes, leaving that patch without CI flag at all.

With this dummy script (and the related job) in that case, this job
will do the dummy review and add the CI flag.

Change-Id: I6d636b5538fa81adc2bc0644bd20d50e2d89f82e
Signed-off-by: David Caro <dcaroest@redhat.com>