packaging: Require version 3.6 of the CLI
Currently the engine .spec file requires the CLI package, but it doesn't
require a explicit version. As a result the CLI won't be updated when
the engine is updated. This isn't a big issue, but may be an
inconvenience for users that expect the latest version of the CLI. This
patch updates the .spec of the engine so that it requires version
3.6.0.0 or newer of the CLI.

Change-Id: Iebdc684e9ad33faa6fdf89984eef4dc6f693e5e1
Bug-Url: https://bugzilla.redhat.com/1252073
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>