restapi: Fix the cardinality storage connection extensions
Currently the storage connection extensions element of the host entity
is declared in the XML schema with an unbounded upper limit. This means
that the corresponding Java method generated by the XJC compiler will
return a list of elements:

  List<StorageConnectionExtensions> getStorageConnectionExtensions() {
   ...
  }

This is incorrect and it causes an issue in the generator of the Java
SDK, as the code generator assumes that this kind of attribute will have
a getter that returns only one element:

  HostStorageConnectionExtensions getStorageConnectionExtensions() {
    ...
  }

The net result is a compilation problem in the generated SDK.

To avoid this issue this patch fixes the XML schema for this element so
that the upper limit is 1 instead of unbounded.

Change-Id: I7f64d74ddc31d5f2634855e31548fa0342e7a281
Bug-Url: https://bugzilla.redhat.com/1254295
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit f41f2b1608bd095d27a2ec71f1479cdd4d71e4c6)