webadmin: edit pool dialog not load when the DC is not up
The problem was that the PoolListModel started the progress by directly calling
startProgress() wich is not good - it has to be handled by the infrastructure
(e.g. AbstractModelBoundPopupPresenterWidget is handling the start/stop
progress according to the AsyncOperation[Started/Completed]Event.

Other fix needed was a null check on the spiceProxyEnabled since the model is
not inited completely when the DC is not active.

Change-Id: I8177155885dfb4c64785b905db614bc6b4ffe44a
Bug-Url: https://bugzilla.redhat.com/1245599
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>