core: Don't fire event if certification weren't retrieved
Reporting 'Failure to retrieve certificates' as an event log
should be skipped to reduce confusion. If there are not certificates,
there is probably communication issue with the host, and logging it
as an error to the log should be sufficient.

Change-Id: I4744ede410ca89ee12a153e43136652ed04dae3e
Bug-Url: https://bugzilla.redhat.com/1253510
Signed-off-by: Moti Asayag <masayag@redhat.com>
(cherry picked from commit 7b2daec6f635f4e9a560ae162168c1ff62488bbc)