core: Fixed Quota description size limits
When trying to copy a quota an error messsage was thrown saying that
the size limit should be between 1 and 250. This error message was the
result of not giving the right limits for the description of the quota.

An empty description is 0 character long. Therefore, the limit was
changed and now it is [0-250].

Change-Id: I6acad4fe290552dcdbd73bcd77e12c65f924d3c2
Bug-Url: https://bugzilla.redhat.com/1245664
Signed-off-by: Tomer Saban <tsaban@redhat.com>