userportal: User portal error message when changing VM cluster inconsistent with webadmin
The error message displayed in user portal when changing
VM cluster is inconsistant with the admin portal. This is
because the user portal invoked first the ChangeVMCluster
command, and then the UpdateVm command, without checking
if the first command was successful. The meaningful message
from the first command was overshaddowed by the less
accurate message from the second command. Now the second
command will only be invoked if the first command was
successful.

Change-Id: I6daaf41c4e09516eb503178b39d3a603dc11c7d1
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1149092
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>