core, restapi: Avoid NPE when there are no provider certificates
Currently the backend and the RESTAPI generate a NPE when there are no
certificates available for a provider, for example if it is using HTTP
instead of HTTPS. The backend generates a log error and then returns
null as the result of the query. The RESTAPI takes this and assumes that
it is a list, thus generating the NPE. This patch changes the backend so
that it returns an empty list in this case and the RESTAPI so that it
tolerates null. The result of trying to import certificates after the
patch will be the error message generated by the backend:

  <fault>
    <reason>Operation Failed</reason>
    <detail>[Failed to import provider certificate.]</detail>
  </fault>

Change-Id: Ie3617dffe56fef88c9267a5471fe41a7ab294254
Bug-Url: https://bugzilla.redhat.com/1255848
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit f6af3c25237e4fc8add4f0f0d6be8f5b43cf6366)