core: Fix, retrieve dedicatedHostList on new VmNumaNode result NPE
The dedicatedHostList is passed to VmNumaNodeCommand in parameters.
Retrieving the VM from parent command result in NPE.
getVm() returns null, because it query on vms view.
The VM is not registered in vms view until it saves its vm_statistics
record and vm_dynamic record as well.
The best practice is to pass and read parameters.

This bug broke the REST api as well.

Change-Id: I4b63a76d9b7e40077dd39605d3aabb0a54b82dcf
Bug-Url: http://bugzilla.redhat.com/1254665
Signed-off-by: Dudi Maroshi <dudi@redhat.com>