restapi: Get provider certificates to import from user
Currently the RESTAPI action used to import provider certificates gets
the certificates to import from the backend and then, in the same
request, it asks the backend to import them. This is not very safe
because there is no guarantee that the caller has ever seen or approved
these certificates. This patch changes the action so that the caller is
forced to send the certificates that are to be approved:

  POST /openstackimageproviders/{provider:id}/importcertificates
  <action>
    <certificates>
      <certificate>the certificate to import</certificate>
    </certificates>
  </action>

The certificates can already be retrieved by the caller as follows:

  GET /openstackimageproviders/{provider:id}/certificates

Change-Id: I8b266fdc6769a84915cb43b9eb97d2f7b4f165b2
Bug-Url: https://bugzilla.redhat.com/1255862
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>