restapi: Locate VM data using pool id, not name
Currently when a VM pool is updated the VM data is located using the
name of the pool. But the name of the pool may be also updated during
the same operation, which may result in not finding the VM data because
there is no such pool yet. To avoid this potential (and real) conflict
this patch changes the RESTAPI so that it uses the identifier of the
pool instead of the name, as the identifier will never change.

Change-Id: I387d122d1019f25f026806ed0042aa32b5db8698
Bug-Url: https://bugzilla.redhat.com/1256411
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 955c08a8ef661305da86a658d3597f7e2940b599)