core: more accurate audit log when domain is deactivated
when a Storage Domain is being deactivated it first moves to "Preparing for
maintenance" status and only when it's confirmed that its not accessed
(by monitoring) from any host it moves to Maintenance status.
However, currently when it moves to "Preparing for maintenance" the
audit log indicates that the domain was deactivated and when it actually
moves to Maintenance there's no audit log at all.
This patch fixes the current audit log and adds a new one for the time
the domain status is changed to Maintenance.

Change-Id: Ia64ce39aeeb4a8874a6c92dc4b337e7e4b1664bc
Bug-Url: https://bugzilla.redhat.com/1247957
Signed-off-by: Liron Aravot <laravot@redhat.com>