core: support domains maintenance on more hosts statuses
When a domain is being deactivated its status is changed to
PreparingForMaintenance. When it's confirmed that no host in the DC access it
(according to the host domain monitoring report) its status is changed
to Maintenance.

The domain monitoring results of hosts in status other then Up status aren't
being analyzed, therfore when there is a host in that status deactivated
domains will remain in "Preparing for maintenance".

An example for a relevant status is NonOperational, As host can be in Non operational
due to various reasons, the domain monitoring results of such hosts should
be anazlyzed as well when inspecting if a domain can be moved to maintenance status
because if the domain isn't accessed it can move safely to
"maintenance".

As part of this change when a domain is deactivated the pool information is
refreshed on hosts in the relevant statuses in dc that uses the memory
backend (to not extend of risk on long connectStoragePool for non
operational hosts as the metadata is stored on the storage). While in DC
that stores the metadata on storage the metadata refresh will occur by
the host recovery mechanism.

The domain monitoring report of those hosts is then inspected as well to allow the
domain to move to maintenance. In this patch I've added only the NonOperational
status as one as the inspected statuses - if it'll be needed we can add
other statuses as well.

Change-Id: I767106d429f0d8b4bd24b0f539b9a91204be18e9
Bug-Url: https://bugzilla.redhat.com/1256841
Signed-off-by: Liron Aravot <laravot@redhat.com>