engine: Fix message when importing SD
When importing a storage domain created on a DC level lower than 3.5, a
log event is displayed.

The content of the event is now :
'This Data center compatibility version does not support importing a
data domain with its entities (VM's and Templates). The imported domain
will be imported without them.'

Change-Id: I490777b044158029dcacd8baf056c2c73a25bb68
Bug-Url: https://bugzilla.redhat.com/1182109
Signed-off-by: Fred Rolland <frolland@redhat.com>