engine: Introduce KeyValueFormatValidation
validation for <key>=<value> spaced separated format.

Please note that extra space is needed in the message ( < key >=< value >) as the message is
being translating to html.
The value between the <> is not being shown.
The usage of &lt &gt was avoided in order to make the language translation easier.

Bug-Url: https://bugzilla.redhat.com/1167812
Change-Id: I78ae0322d49356ba0ef9e0e496a232348a715eef
Signed-off-by: Eliraz Levi <elevi@redhat.com>