restapi: Treat template watchdogs as template devices
When support for WildFly was intruduced a series of changes were needed
to move the DELETE methods from collections to resources. The change to
the watchdogs resources introduced a regression in the treatment of
template watchdogs: they are treated as VM resources instead of template
resources. This patch fixes that issue.

Change-Id: I8c8dbad721446e71bd6dc135af37ad42fd8518bf
Bug-Url: https://bugzilla.redhat.com/1258224
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>