restapi: Remove @Path from storage connection extension implementation
Currently the implementation of the connection extension resource
contains a @Path annotation. This annotation is incorrect, as this kind
of annotation should go only in the interface, not in the
implementation. In addition the annotation value is "/whatever" and that
is incorrect as well because this isn't a top level resource. This
incorrect annotation induces an error in the version of Resteasy
included in JBoss EAP. This patch removes the annotation.

Change-Id: Ia3e7b8876417e2437db8f1ddd55ea633d32ae918
Bug-Url: https://bugzilla.redhat.com/1258247
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>