core: add query to support restapi
RESTAPI will expose ssh public key as collection - not as single
instance, to be future compatible.

However, backend knows there is at most one key, embedded inside the
UserProfile.

This patch fills the gap, returning a list of UserProfiles to fit
restapi needs.

The most correct approach would be to return a list of backend'
SSHPublicKeys, but backend has not this Class yet.

Change-Id: Ie39404e9eccb15480f79e4438489b8aa0e0895df
Bug-Url: https://bugzilla.redhat.com/1254445
Signed-off-by: Francesco Romani <fromani@redhat.com>