core: improve balance selection of vm for migration
Iterate over source hosts, to find vm that can be migrated,
without over utilize memory or CPU of destination host.

Before was selected some random source host, and from this host
was selected single vm with minimum cpu usage,
but maybe with big amount of memory usage and
so over utilizied destination host memory.

In result we had that balancing not migrate any of vms.

Change-Id: I67947fa4463b8bb12d04ee8f4126fa93683c9e14
Bug-Url: https://bugzilla.redhat.com/1258780
Signed-off-by: Artyom Lukianov <alukiano@redhat.com>