utils: allow spaces within encrypted fields
Ever since 3.0 (probably even before) the encryption of database fields
that was used is invalid, it uses RSA to encrypt blobs instead of using
ciphers within envelope.

It also stores null and empty strings as plain, and to make it even
better it trims spaces out of the input for some reason.

To conclude security... if decryption fails it falls back to use the
blob as plain text.

This logic was untouched, under the hope that we slowly remove usages of
it.

AAA does not use it any more, we should remove all.

For now, we remove the trim() as if the password of trim() actually
works so far it will keep working, new passwords with leading/trailing
spaces will be rejected.

The risk is if for some reason we have " "* in database field it will be
rejected as valid password, fixing it will be re-set password by user to
empty one.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1258867
Change-Id: I2ca15519ec245efd82e71f2ec39abd4ca1fe81c2
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>