core: nesting PluralMessages into NetworkValidator
PluralMessages was intended to help with singular-plural localization
pairs. This is bad approach, but I thought it should help with
transition. But when variable replacements hit this class it becomes
really unreadable and complexity overcame the benefits.

This class was also used from NetworkAttachmentValidator, but that one
can delegate to NetworkValidator, so PluralMessages was simplified a
lot, uses newly available methods, and is made accessible only to
NetworkValidator.

Change-Id: I6c88ed8f66ace2c515d1174cf305204074f432ce
Signed-off-by: Martin Mucha <mmucha@redhat.com>