restapi: Introduce metamodel
Currently the model of the RESTAPI is defined using a XML schema. This
XML schema is processed with the JAXB "xjc" compiler and the Java
classes of the model are generated. This has several inconveniences:

1. The model is tied to the XML schema model, even if we aim to
   support both XML and JSON. JSON happens to work correctly because
   we are lucky enough that the JSON library that we use supports the
   JAXB annotations.

2. It is difficult to extract information from the XML schema. Tools
   like the generators of the SDKs have to do some ugly tricks, like
   using XPath queries and using reflection to ispect the code
   generated by the JAXB "xjc" compiler.

3. The XML schema is single giant file, more than 6000 lines at the
   moment.

In order to improve the situation this patch introduces a metamodel that
will be used to describe the data types used by the RESTAPI. This
metamodel will be written in Java, using interfaces and annotations to
describe the types. For example, to describe the bookmark type an
interface like the following will be used:

  @Type
  public interface Bookmark {
    String value();
  }

This only purpose of this interface is to describe the characteristics
of the type: its name, its attributes, etc. It isn't intended to be
implemented or used during runtime in any way. The metamodel analyzer
will inspect this source code and create a data structure easier to
digest and use for applications, like the generators of the SDKs. For
example, a tool could use the generated data structure to list the
attributes of the bookmark type as follows:

  // Analyze the model files:
  Model model = new Model();
  ModelAnalyzer analyzer = new ModelAnalyzer();
  analyzer.setModel(model);
  analyzer.analyzeSource(...);

  // Find the bookmark type and list its attributes:
  Name bookmarkName = NameParser.parseUsingCase("Bookmark");
  Type bookmarType = model.getType(bookmarkName);
  for (Attribute attribute : bookmarkType.getAttributes()) {
    System.out.println(attribute.getName());
  }

The metamodel includes also a tool that will be used to generate code
from the model. Currently it only supports generating fragmetns of the
XML schema. This is needed in order to preserve compatibility with the
current RESTAPI implementation. The tool will take the existing XML
schema, and will add to it the fragments generated to the model. This
way it is possible to gradually replace the XML schema description of
types with the equivalent Java description.

In the future the tool will be extended to generate other outputs:

* Java interfaces, containers and builders to be used during runtime,
  instead of the code currently generated by the JAXB "xjc" compiler.

* Java code used to parse and serialize XML and JSON documents using an
  streaming approach, instead of the "all in memory" approach used by JAXB
  (and also by the JSON implementation that we use).

* Documentation, extracted from the Javadoc comments of the type
  interfaces.

Change-Id: I80b036b5c3b3f9ad4d316c8f6c0a568b0465adb9
Related: https://bugzilla.redhat.com/1201197
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>