engine: GetDeviceList without check status
When a user opens the Storage Domain or Disk dialog for block storage,
a GetDeviceList verb is called on VDSM.
In order to populate the 'status' field, the getDeviceList verb perform
a create PV test. This operation is expensive, and in setups with large
number of devices it will cause performance problem and even timeouts.

The flow before this patch was:
    - Engine calls getDeviceList, VDSM performs status check for each
      device, possibly times out
    - User selects a few devices (typically one) to add/extend a SD
      or create a new LUN disk
    - GUI warns user about used devices

The flow with this patch:
    - Engine calls getDeviceList, skipping status check
    - User selects a few devices (typically one) to add/extend a SD
      or create a new LUN disk
    - Engine calls getDeviceList, checking status only for selected
    - Engine warns user about used devices

Change-Id: Ib4484307e3d4ac66a4370bcef3ac015ad0f2d084
Bug-Url: https://bugzilla.redhat.com/1217401
Signed-off-by: Fred Rolland <frolland@redhat.com>