core: Fix checking compatibility of proxy host and fencing policy
Fixes the issue when checking compatibility of fence proxy host with
fencing policy caused by BZ1229177. It changes the comparison using
following rule: If at least one version contained in host supported
cluster levels is higher or equals to minimal required version for the
fencing policy of the cluster, then host can be used as a proxy.

Change-Id: I94cc1419186dfd3ab557143e41b3d3f3772f614a
Bug-Url: https://bugzilla.redhat.com/1257610
Signed-off-by: Martin Perina <mperina@redhat.com>