core: Avoid NPEs on illegal LUN disk operations
Move, copy and export of LUN disks, either attached to a VM or not is not
possible. Using REST, lack of image id and source storage domain resulted in
NPEs. These are now checked before being sent to the engine, while the
appropriate commands check that the disk is not of type LUN (which
applies for all above null scenarios).

Change-Id: I6de6615cd8f154b9c348055e65d153c9da0f12e8
Bug-Url: https://bugzilla.redhat.com/1147861
Bug-Url: https://bugzilla.redhat.com/1147863
Bug-Url: https://bugzilla.redhat.com/1147870
Bug-Url: https://bugzilla.redhat.com/1147873
Bug-Url: https://bugzilla.redhat.com/1147874
Bug-Url: https://bugzilla.redhat.com/1126425
Bug-Url: https://bugzilla.redhat.com/1147860
Signed-off-by: Vered Volansky <vvolansk@redhat.com>