core: fix events handling after host reboot
The engine filters events if it processed newer events already.
It compares the timestamp on each events with the timestamp of
the latest event that came from the same host and has been
processed.

VDSM uses monotonic time for the timestamp on the status
events. The engine relay on the timestamp being monotonic.
However, the timestamp is not monotonic after the host reboots.

The solution is to clear the last host-id that the latest event
that has been processed came from if that host has been
rebooted. That way, we won't use timestamp that has been
received before the host has been rebooted and the next event
will be processed.

Change-Id: Iba99b28312e1b729cd48dda80bac464f32ba7828
Bug-Url: https://bugzilla.redhat.com/1255033
Signed-off-by: Arik Hadas <ahadas@redhat.com>